@@ -1,22 +1,20 @@
 name: mobility_features_example
 description: An example app for the Mobility Features package.
-publish_to: "none" # Remove this line if you wish to publish to pub.dev
-version: 3.1.0
+publish_to: "none" 
 
 environment:
-  sdk: ">=2.17.0 <4.0.0"
+  sdk: ">=3.2.0 <4.0.0"
+  flutter: ">=3.16.0"
 
 dependencies:
   flutter:
     sdk: flutter
+  permission_handler: ^11.3.1
+  carp_background_location: ^4.0.0
 
   mobility_features:
     path: ../
-  carp_background_location:
-    ^4.0.0
-    # path: ../../carp_background_location
 
-  cupertino_icons: ^1.0.5
 
 dev_dependencies:
   flutter_test:
