@@ -1,140 +1,147 @@
-part of mobility_features;
+part of '../mobility_features.dart';
 
-/// Daily mobility context.
+/// Daily mobility context for a day on [date].
 ///
-/// All Stops and Moves are on the same [date].
-/// [places] are all places for which the duration on the given date is greater than 0.
+/// All [stops] and [moves] are on the same [date].
+/// The [places] are all places for which the duration on the given [date]
+/// is greater than 0.
+@JsonSerializable(includeIfNull: false, explicitToJson: true)
 class MobilityContext {
-  late DateTime _timestamp, _date;
-  List<Stop> _stops;
-  List<Place> _places;
-  List<Move> _moves;
-  late List<Place> _significantPlaces;
+  List<Place>? _significantPlaces;
+  HourMatrix? _hourMatrix;
   Place? _homePlace;
-
-  late _HourMatrix _hourMatrix;
-
-  double? _locationVariance,
-      _entropy,
-      _normalizedEntropy,
-      _homeStay,
-      _distanceTravelled;
-  List<MobilityContext>? contexts;
-
-  /// Private constructor, cannot be instantiated from outside
-  MobilityContext._(this._stops, this._places, this._moves, this._date) {
-    _timestamp = DateTime.now();
-
-    // if contexts array is null, init to empty array
-    contexts = contexts ?? [];
-
-    // compute all the features
-    _significantPlaces =
-        _places.where((p) => p.duration > Duration(minutes: 3)).toList();
-    _hourMatrix = _HourMatrix.fromStops(_stops, _places.length);
-    _homePlace = _findHomePlaceToday();
-    _homeStay = _calculateHomeStay();
-    _locationVariance = _calculateLocationVariance();
-    _entropy = _calculateEntropy();
-    _normalizedEntropy = _calculateNormalizedEntropy();
-    _distanceTravelled = _calculateDistanceTravelled();
-  }
-
-  // The date of this context.
-  DateTime get date => _date;
+  List<Stop>? _stops;
+  List<Place>? _places;
+  List<Move>? _moves;
 
   /// Timestamp at which the features were computed
-  DateTime get timestamp => _timestamp;
+  DateTime? timestamp;
 
-  /// Stops today.
-  List<Stop> get stops => _stops;
+  /// The date of this context.
+  DateTime? date;
 
-  /// Moves today.
-  List<Move> get moves => _moves;
+  /// Number of stops made today.
+  int? numberOfStops;
 
-  /// Places today.
-  List<Place> get places => _places;
-
-  /// All significant places, i.e. places with a minimum stay duration.
-  List<Place> get significantPlaces => _significantPlaces;
+  /// Number of moves made today.
+  int? numberOfMoves;
 
   /// Number of significant places visited today.
-  int get numberOfSignificantPlaces => _significantPlaces.length;
-
-  /// Home place.
-  /// Returns null if home cannot be found from the available data.
-  Place? get homePlace => _homePlace;
-
-  /// Home Stay Percentage today. A scalar between 0 and 1.
-  /// Returns null if cannot be calculated based on the available data.
-  double? get homeStay => _homeStay;
+  int? numberOfSignificantPlaces;
 
   /// Location variance today.
-  double? get locationVariance => _locationVariance;
+  double? locationVariance;
 
   /// Location entropy.
   ///
   ///  * High entropy: Time is spent evenly among all places
   ///  * Low  entropy: Time is mainly spent at a few of the places
-  double? get entropy => _entropy;
+  double? entropy;
 
   /// Normalized location entropy. A scalar between 0 and 1.
-  double? get normalizedEntropy => _normalizedEntropy;
+  double? normalizedEntropy;
 
-  /// Distance travelled today in meters.
-  double? get distanceTravelled => _distanceTravelled;
+  /// Home Stay Percentage today. A scalar between 0 and 1.
+  /// Returns null if cannot be calculated based on the available data.
+  double? homeStay;
 
-  /// Private home stay calculation
+  /// Distance traveled today in meters.
+  double? distanceTraveled;
+
+  MobilityContext();
+
+  MobilityContext.fromMobility(
+    this.date,
+    this._stops,
+    this._places,
+    this._moves,
+  ) {
+    timestamp = DateTime.now();
+
+    // compute all features
+    _significantPlaces =
+        _places!.where((p) => p.duration > const Duration(minutes: 3)).toList();
+    _hourMatrix = HourMatrix.fromStops(_stops!, _places!.length);
+    numberOfSignificantPlaces = _significantPlaces!.length;
+    numberOfStops = _stops?.length;
+    numberOfMoves = _moves?.length;
+    _homePlace = _findHomePlaceToday();
+    homeStay = _calculateHomeStay();
+    locationVariance = _calculateLocationVariance();
+    entropy = _calculateEntropy();
+    normalizedEntropy = _calculateNormalizedEntropy();
+    distanceTraveled = _calculateDistanceTraveled();
+  }
+
+  /// Stops today.
+  List<Stop>? get stops => _stops;
+
+  /// Places today.
+  List<Place>? get places => _places;
+
+  /// Moves today.
+  List<Move>? get moves => _moves;
+
+  /// The place used as 'home'.
+  /// Returns null if home cannot be found from the available data.
+  Place? get homePlace => _homePlace;
+
+  /// All significant places, i.e. places with a minimum stay duration.
+  List<Place>? get significantPlaces => _significantPlaces;
+
   double? _calculateHomeStay() {
-    if (stops.isEmpty) return null;
+    if (_stops == null) return null;
+    if (_stops!.isEmpty) return null;
 
     // Latest known sample time
-    DateTime latestTime = _stops.last.departure;
+    final latestTime = _stops!.last.departure;
 
     // Total time elapsed from midnight until the last stop
     int totalTime = latestTime.millisecondsSinceEpoch -
         latestTime.midnight.millisecondsSinceEpoch;
 
     // Find todays home id, if no home exists today return null
-    if (_hourMatrix.homePlaceId == -1) return null;
+    if (_hourMatrix!.homePlaceId == -1) return null;
 
-    int homeTime = stops
-        .where((s) => s.placeId == _hourMatrix.homePlaceId)
+    int homeTime = _stops!
+        .where((s) => s.placeId == _hourMatrix!.homePlaceId)
         .map((s) => s.duration.inMilliseconds)
         .fold(0, (a, b) => a + b);
 
     return homeTime.toDouble() / totalTime.toDouble();
   }
 
-  Place? _findHomePlaceToday() => (_hourMatrix.homePlaceId == -1)
+  Place? _findHomePlaceToday() => (_hourMatrix!.homePlaceId == -1)
       ? null
-      : _places.where((p) => p.id == _hourMatrix.homePlaceId).first;
+      : _places!.where((p) => p.id == _hourMatrix!.homePlaceId).first;
 
-  /// Location variance calculation
   double? _calculateLocationVariance() {
     // Require at least 2 observations
-    if (_stops.length < 2) return 0.0;
+    if (_stops!.length < 2) return 0.0;
 
-    double latStd = Stats.fromData(_stops.map((s) => (s.geoLocation.latitude)))
+    double latStd = Stats.fromData(_stops!.map((s) => (s.geoLocation.latitude)))
         .standardDeviation as double;
 
-    double lonStd = Stats.fromData(_stops.map((s) => (s.geoLocation.longitude)))
-        .standardDeviation as double;
+    double lonStd =
+        Stats.fromData(_stops!.map((s) => (s.geoLocation.longitude)))
+            .standardDeviation as double;
     return log(latStd * latStd + lonStd * lonStd + 1);
   }
 
   double? _calculateEntropy() {
     // if no places were visited return null
-    if (places.isEmpty)
+    // else - the Entropy is zero when one outcome is certain to occur
+    if (_places!.isEmpty) {
       return null;
-    // the Entropy is zero when one outcome is certain to occur
-    else if (places.length == 1) return 0.0;
+    } else if (_places!.length == 1) {
+      return 0.0;
+    }
 
     // calculate time spent at different places
     List<Duration> durations =
-        places.map((p) => p.durationForDate(date)).toList();
+        _places!.map((p) => p.durationForDate(date)).toList();
 
-    Duration totalTimeSpent = durations.fold(Duration(), (a, b) => a + b);
+    Duration totalTimeSpent = durations.fold(const Duration(), (a, b) => a + b);
 
     List<double> distribution = durations
         .map((d) => (d.inMilliseconds.toDouble() /
@@ -144,25 +151,15 @@ class MobilityContext {
     return -distribution.map((p) => p * log(p)).reduce((a, b) => (a + b));
   }
 
-  /// Private normalized entropy calculation
   double _calculateNormalizedEntropy() =>
-      (places.length == 1) ? 0.0 : entropy! / log(places.length);
+      (_places!.length == 1) ? 0.0 : entropy! / log(_places!.length);
 
-  /// Private distance travelled calculation
-  double _calculateDistanceTravelled() =>
-      _moves.map((m) => (m.distance)).fold(0.0, (a, b) => a + b!);
+  double _calculateDistanceTraveled() =>
+      _moves!.map((m) => (m.distance)).fold(0.0, (a, b) => a + b!);
 
-  Map<String, dynamic> toJson() => {
-        "date": date.toIso8601String(),
-        "computed_at": timestamp.toIso8601String(),
-        "num_of_stops": stops.length,
-        "num_of_moves": moves.length,
-        "num_of_significant_places": significantPlaces.length,
-        "normalized_entropy": normalizedEntropy,
-        "home_stay": homeStay,
-        "distance_travelled": distanceTravelled,
-        "location_variance": locationVariance
-      };
+  factory MobilityContext.fromJson(Map<String, dynamic> json) =>
+      _$MobilityContextFromJson(json);
+  Map<String, dynamic> toJson() => _$MobilityContextToJson(this);
 }
 
 //  /// Routine index (overlap) calculation
