@@ -1,13 +1,16 @@
 part of '../health.dart';
 
 /// Types of health platforms.
-enum HealthPlatformType { appleHealth, googleFit, googleHealthConnect }
+enum HealthPlatformType { appleHealth, googleHealthConnect }
 
 /// A [HealthDataPoint] object corresponds to a data point capture from
-/// Apple HealthKit or Google Fit or Google Health Connect with a [HealthValue]
+/// Apple HealthKit or Google Health Connect with a [HealthValue]
 /// as value.
-@JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
+@JsonSerializable(includeIfNull: false, explicitToJson: true)
 class HealthDataPoint {
+  /// UUID of the data point.
+  String uuid;
+
   /// The quantity value of the data point
   HealthValue value;
 
@@ -41,13 +44,19 @@ class HealthDataPoint {
   /// The name of the source from which the data point was fetched.
   String sourceName;
 
-  /// The user entered state of the data point.
-  bool isManualEntry;
+  /// How the data point was recorded
+  /// (on Android: https://developer.android.com/reference/kotlin/androidx/health/connect/client/records/metadata/Metadata#summary)
+  /// on iOS: either user entered or manual https://developer.apple.com/documentation/healthkit/hkmetadatakeywasuserentered)
+  RecordingMethod recordingMethod;
 
   /// The summary of the workout data point, if available.
   WorkoutSummary? workoutSummary;
 
+  /// The metadata for this data point.
+  Map<String, dynamic>? metadata;
+
   HealthDataPoint({
+    required this.uuid,
     required this.value,
     required this.type,
     required this.unit,
@@ -57,8 +66,9 @@ class HealthDataPoint {
     required this.sourceDeviceId,
     required this.sourceId,
     required this.sourceName,
-    this.isManualEntry = false,
+    this.recordingMethod = RecordingMethod.unknown,
     this.workoutSummary,
+    this.metadata,
   }) {
     // set the value to minutes rather than the category
     // returned by the native API
@@ -68,12 +78,14 @@ class HealthDataPoint {
         type == HealthDataType.HEADACHE_MILD ||
         type == HealthDataType.HEADACHE_MODERATE ||
         type == HealthDataType.HEADACHE_SEVERE ||
-        type == HealthDataType.SLEEP_IN_BED ||
         type == HealthDataType.SLEEP_ASLEEP ||
         type == HealthDataType.SLEEP_AWAKE ||
+        type == HealthDataType.SLEEP_AWAKE_IN_BED ||
         type == HealthDataType.SLEEP_DEEP ||
+        type == HealthDataType.SLEEP_IN_BED ||
         type == HealthDataType.SLEEP_LIGHT ||
         type == HealthDataType.SLEEP_REM ||
+        type == HealthDataType.SLEEP_UNKNOWN ||
         type == HealthDataType.SLEEP_OUT_OF_BED) {
       value = _convertMinutes();
     }
@@ -107,6 +119,10 @@ class HealthDataPoint {
         ElectrocardiogramHealthValue.fromHealthDataPoint(dataPoint),
       HealthDataType.NUTRITION =>
         NutritionHealthValue.fromHealthDataPoint(dataPoint),
+      HealthDataType.INSULIN_DELIVERY =>
+        InsulinDeliveryHealthValue.fromHealthDataPoint(dataPoint),
+      HealthDataType.MENSTRUATION_FLOW =>
+        MenstruationFlowHealthValue.fromHealthDataPoint(dataPoint),
       _ => NumericHealthValue.fromHealthDataPoint(dataPoint),
     };
 
@@ -116,8 +132,11 @@ class HealthDataPoint {
         DateTime.fromMillisecondsSinceEpoch(dataPoint['date_to'] as int);
     final String sourceId = dataPoint["source_id"] as String;
     final String sourceName = dataPoint["source_name"] as String;
-    final bool isManualEntry = dataPoint["is_manual_entry"] as bool? ?? false;
+    final Map<String, dynamic>? metadata = dataPoint["metadata"] == null
+        ? null
+        : Map<String, dynamic>.from(dataPoint['metadata'] as Map);
     final unit = dataTypeToUnit[dataType] ?? HealthDataUnit.UNKNOWN_UNIT;
+    final String? uuid = dataPoint["uuid"] as String?;
 
     // Set WorkoutSummary, if available.
     WorkoutSummary? workoutSummary;
@@ -128,7 +147,10 @@ class HealthDataPoint {
       workoutSummary = WorkoutSummary.fromHealthDataPoint(dataPoint);
     }
 
+    var recordingMethod = dataPoint["recording_method"] as int?;
+
     return HealthDataPoint(
+      uuid: uuid ?? "",
       value: value,
       type: dataType,
       unit: unit,
@@ -138,13 +160,15 @@ class HealthDataPoint {
       sourceDeviceId: Health().deviceId,
       sourceId: sourceId,
       sourceName: sourceName,
-      isManualEntry: isManualEntry,
+      recordingMethod: RecordingMethod.fromInt(recordingMethod),
       workoutSummary: workoutSummary,
+      metadata: metadata,
     );
   }
 
   @override
   String toString() => """$runtimeType -
+    uuid: $uuid,
     value: ${value.toString()},
     unit: ${unit.name},
     dateFrom: $dateFrom,
@@ -154,12 +178,14 @@ class HealthDataPoint {
     deviceId: $sourceDeviceId,
     sourceId: $sourceId,
     sourceName: $sourceName
-    isManualEntry: $isManualEntry
-    workoutSummary: $workoutSummary""";
+    recordingMethod: $recordingMethod
+    workoutSummary: $workoutSummary
+    metadata: $metadata""";
 
   @override
   bool operator ==(Object other) =>
       other is HealthDataPoint &&
+      uuid == other.uuid &&
       value == other.value &&
       unit == other.unit &&
       dateFrom == other.dateFrom &&
@@ -169,9 +195,10 @@ class HealthDataPoint {
       sourceDeviceId == other.sourceDeviceId &&
       sourceId == other.sourceId &&
       sourceName == other.sourceName &&
-      isManualEntry == other.isManualEntry;
+      recordingMethod == other.recordingMethod &&
+      metadata == other.metadata;
 
   @override
-  int get hashCode => Object.hash(value, unit, dateFrom, dateTo, type,
-      sourcePlatform, sourceDeviceId, sourceId, sourceName);
+  int get hashCode => Object.hash(uuid, value, unit, dateFrom, dateTo, type,
+      sourcePlatform, sourceDeviceId, sourceId, sourceName, metadata);
 }
