@@ -1,43 +1,123 @@
 part of '../health.dart';
 
-/// List of data types available on Android
-const List<HealthDataType> dataTypeKeysAndroid = [
-  HealthDataType.ACTIVE_ENERGY_BURNED,
-  HealthDataType.BLOOD_GLUCOSE,
-  HealthDataType.BLOOD_OXYGEN,
-  HealthDataType.BLOOD_PRESSURE_DIASTOLIC,
-  HealthDataType.BLOOD_PRESSURE_SYSTOLIC,
-  HealthDataType.BODY_FAT_PERCENTAGE,
-  HealthDataType.BODY_MASS_INDEX,
-  HealthDataType.BODY_TEMPERATURE,
-  HealthDataType.BODY_WATER_MASS,
-  HealthDataType.HEART_RATE,
-  HealthDataType.HEIGHT,
-  HealthDataType.STEPS,
-  HealthDataType.WEIGHT,
-  HealthDataType.MOVE_MINUTES,
-  HealthDataType.DISTANCE_DELTA,
-  HealthDataType.SLEEP_AWAKE,
-  HealthDataType.SLEEP_ASLEEP,
-  HealthDataType.SLEEP_IN_BED,
-  HealthDataType.SLEEP_DEEP,
-  HealthDataType.SLEEP_LIGHT,
-  HealthDataType.SLEEP_REM,
-  HealthDataType.SLEEP_OUT_OF_BED,
-  HealthDataType.SLEEP_SESSION,
-  HealthDataType.WATER,
-  HealthDataType.WORKOUT,
-  HealthDataType.RESTING_HEART_RATE,
-  HealthDataType.FLIGHTS_CLIMBED,
-  HealthDataType.BASAL_ENERGY_BURNED,
-  HealthDataType.RESPIRATORY_RATE,
-  HealthDataType.NUTRITION,
-  HealthDataType.TOTAL_CALORIES_BURNED,
-];
+/// List of all available health data types.
+enum HealthDataType {
+  ACTIVE_ENERGY_BURNED,
+  ATRIAL_FIBRILLATION_BURDEN,
+  AUDIOGRAM,
+  BASAL_ENERGY_BURNED,
+  BLOOD_GLUCOSE,
+  BLOOD_OXYGEN,
+  BLOOD_PRESSURE_DIASTOLIC,
+  BLOOD_PRESSURE_SYSTOLIC,
+  BODY_FAT_PERCENTAGE,
+  BODY_MASS_INDEX,
+  BODY_TEMPERATURE,
+  BODY_WATER_MASS,
+  DIETARY_CARBS_CONSUMED,
+  DIETARY_CAFFEINE,
+  DIETARY_ENERGY_CONSUMED,
+  DIETARY_FATS_CONSUMED,
+  DIETARY_PROTEIN_CONSUMED,
+  DIETARY_FIBER,
+  DIETARY_SUGAR,
+  DIETARY_FAT_MONOUNSATURATED,
+  DIETARY_FAT_POLYUNSATURATED,
+  DIETARY_FAT_SATURATED,
+  DIETARY_CHOLESTEROL,
+  DIETARY_VITAMIN_A,
+  DIETARY_THIAMIN,
+  DIETARY_RIBOFLAVIN,
+  DIETARY_NIACIN,
+  DIETARY_PANTOTHENIC_ACID,
+  DIETARY_VITAMIN_B6,
+  DIETARY_BIOTIN,
+  DIETARY_VITAMIN_B12,
+  DIETARY_VITAMIN_C,
+  DIETARY_VITAMIN_D,
+  DIETARY_VITAMIN_E,
+  DIETARY_VITAMIN_K,
+  DIETARY_FOLATE,
+  DIETARY_CALCIUM,
+  DIETARY_CHLORIDE,
+  DIETARY_IRON,
+  DIETARY_MAGNESIUM,
+  DIETARY_PHOSPHORUS,
+  DIETARY_POTASSIUM,
+  DIETARY_SODIUM,
+  DIETARY_ZINC,
+  DIETARY_CHROMIUM,
+  DIETARY_COPPER,
+  DIETARY_IODINE,
+  DIETARY_MANGANESE,
+  DIETARY_MOLYBDENUM,
+  DIETARY_SELENIUM,
+  FORCED_EXPIRATORY_VOLUME,
+  HEART_RATE,
+  HEART_RATE_VARIABILITY_SDNN,
+  HEART_RATE_VARIABILITY_RMSSD,
+  HEIGHT,
+  INSULIN_DELIVERY,
+  RESTING_HEART_RATE,
+  RESPIRATORY_RATE,
+  PERIPHERAL_PERFUSION_INDEX,
+  STEPS,
+  WAIST_CIRCUMFERENCE,
+  WALKING_HEART_RATE,
+  WEIGHT,
+  DISTANCE_WALKING_RUNNING,
+  DISTANCE_SWIMMING,
+  DISTANCE_CYCLING,
+  FLIGHTS_CLIMBED,
+  DISTANCE_DELTA,
+  MINDFULNESS,
+  WATER,
+  SLEEP_ASLEEP,
+  SLEEP_AWAKE_IN_BED,
+  SLEEP_AWAKE,
+  SLEEP_DEEP,
+  SLEEP_IN_BED,
+  SLEEP_LIGHT,
+  SLEEP_OUT_OF_BED,
+  SLEEP_REM,
+  SLEEP_SESSION,
+  SLEEP_UNKNOWN,
+  EXERCISE_TIME,
+  WORKOUT,
+  HEADACHE_NOT_PRESENT,
+  HEADACHE_MILD,
+  HEADACHE_MODERATE,
+  HEADACHE_SEVERE,
+  HEADACHE_UNSPECIFIED,
+  NUTRITION,
+  // HealthKit Characteristics
+  GENDER,
+  BIRTH_DATE,
+  BLOOD_TYPE,
+  MENSTRUATION_FLOW,
+
+  // Heart Rate events (specific to Apple Watch)
+  HIGH_HEART_RATE_EVENT,
+  LOW_HEART_RATE_EVENT,
+  IRREGULAR_HEART_RATE_EVENT,
+  ELECTRODERMAL_ACTIVITY,
+  ELECTROCARDIOGRAM,
+
+  // Health Connect
+  TOTAL_CALORIES_BURNED
+}
+
+/// Access types for Health Data.
+enum HealthDataAccess {
+  READ,
+  WRITE,
+  READ_WRITE,
+}
 
 /// List of data types available on iOS.
 const List<HealthDataType> dataTypeKeysIOS = [
   HealthDataType.ACTIVE_ENERGY_BURNED,
+  HealthDataType.ATRIAL_FIBRILLATION_BURDEN,
   HealthDataType.AUDIOGRAM,
   HealthDataType.BASAL_ENERGY_BURNED,
   HealthDataType.BLOOD_GLUCOSE,
@@ -52,11 +132,45 @@ const List<HealthDataType> dataTypeKeysIOS = [
   HealthDataType.DIETARY_ENERGY_CONSUMED,
   HealthDataType.DIETARY_FATS_CONSUMED,
   HealthDataType.DIETARY_PROTEIN_CONSUMED,
+  HealthDataType.DIETARY_FIBER,
+  HealthDataType.DIETARY_SUGAR,
+  HealthDataType.DIETARY_FAT_MONOUNSATURATED,
+  HealthDataType.DIETARY_FAT_POLYUNSATURATED,
+  HealthDataType.DIETARY_FAT_SATURATED,
+  HealthDataType.DIETARY_CHOLESTEROL,
+  HealthDataType.DIETARY_VITAMIN_A,
+  HealthDataType.DIETARY_THIAMIN,
+  HealthDataType.DIETARY_RIBOFLAVIN,
+  HealthDataType.DIETARY_NIACIN,
+  HealthDataType.DIETARY_PANTOTHENIC_ACID,
+  HealthDataType.DIETARY_VITAMIN_B6,
+  HealthDataType.DIETARY_BIOTIN,
+  HealthDataType.DIETARY_VITAMIN_B12,
+  HealthDataType.DIETARY_VITAMIN_C,
+  HealthDataType.DIETARY_VITAMIN_D,
+  HealthDataType.DIETARY_VITAMIN_E,
+  HealthDataType.DIETARY_VITAMIN_K,
+  HealthDataType.DIETARY_FOLATE,
+  HealthDataType.DIETARY_CALCIUM,
+  HealthDataType.DIETARY_CHLORIDE,
+  HealthDataType.DIETARY_IRON,
+  HealthDataType.DIETARY_MAGNESIUM,
+  HealthDataType.DIETARY_PHOSPHORUS,
+  HealthDataType.DIETARY_POTASSIUM,
+  HealthDataType.DIETARY_SODIUM,
+  HealthDataType.DIETARY_ZINC,
+  HealthDataType.DIETARY_CHROMIUM,
+  HealthDataType.DIETARY_COPPER,
+  HealthDataType.DIETARY_IODINE,
+  HealthDataType.DIETARY_MANGANESE,
+  HealthDataType.DIETARY_MOLYBDENUM,
+  HealthDataType.DIETARY_SELENIUM,
   HealthDataType.ELECTRODERMAL_ACTIVITY,
   HealthDataType.FORCED_EXPIRATORY_VOLUME,
   HealthDataType.HEART_RATE,
   HealthDataType.HEART_RATE_VARIABILITY_SDNN,
   HealthDataType.HEIGHT,
+  HealthDataType.INSULIN_DELIVERY,
   HealthDataType.HIGH_HEART_RATE_EVENT,
   HealthDataType.IRREGULAR_HEART_RATE_EVENT,
   HealthDataType.LOW_HEART_RATE_EVENT,
@@ -72,14 +186,12 @@ const List<HealthDataType> dataTypeKeysIOS = [
   HealthDataType.DISTANCE_SWIMMING,
   HealthDataType.DISTANCE_CYCLING,
   HealthDataType.MINDFULNESS,
-  HealthDataType.SLEEP_IN_BED,
-  HealthDataType.SLEEP_AWAKE,
   HealthDataType.SLEEP_ASLEEP,
+  HealthDataType.SLEEP_AWAKE,
   HealthDataType.SLEEP_DEEP,
+  HealthDataType.SLEEP_IN_BED,
+  HealthDataType.SLEEP_LIGHT,
   HealthDataType.SLEEP_REM,
-  HealthDataType.SLEEP_ASLEEP_CORE,
-  HealthDataType.SLEEP_ASLEEP_DEEP,
-  HealthDataType.SLEEP_ASLEEP_REM,
   HealthDataType.WATER,
   HealthDataType.EXERCISE_TIME,
   HealthDataType.WORKOUT,
@@ -90,6 +202,11 @@ const List<HealthDataType> dataTypeKeysIOS = [
   HealthDataType.HEADACHE_UNSPECIFIED,
   HealthDataType.ELECTROCARDIOGRAM,
   HealthDataType.NUTRITION,
+  HealthDataType.GENDER,
+  HealthDataType.BIRTH_DATE,
+  HealthDataType.BLOOD_TYPE,
+  HealthDataType.MENSTRUATION_FLOW,
+
   // JB: Added
   HealthDataType.STAIR_ASCENT_SPEED,
   HealthDataType.STAIR_DESCENT_SPEED,
@@ -99,9 +216,47 @@ const List<HealthDataType> dataTypeKeysIOS = [
   HealthDataType.WALKING_SPEED,
 ];
 
+/// List of data types available on Android
+const List<HealthDataType> dataTypeKeysAndroid = [
+  HealthDataType.ACTIVE_ENERGY_BURNED,
+  HealthDataType.BLOOD_GLUCOSE,
+  HealthDataType.BLOOD_OXYGEN,
+  HealthDataType.BLOOD_PRESSURE_DIASTOLIC,
+  HealthDataType.BLOOD_PRESSURE_SYSTOLIC,
+  HealthDataType.BODY_FAT_PERCENTAGE,
+  HealthDataType.BODY_MASS_INDEX,
+  HealthDataType.BODY_TEMPERATURE,
+  HealthDataType.BODY_WATER_MASS,
+  HealthDataType.HEART_RATE,
+  HealthDataType.HEART_RATE_VARIABILITY_RMSSD,
+  HealthDataType.HEIGHT,
+  HealthDataType.STEPS,
+  HealthDataType.WEIGHT,
+  HealthDataType.DISTANCE_DELTA,
+  HealthDataType.SLEEP_ASLEEP,
+  HealthDataType.SLEEP_AWAKE_IN_BED,
+  HealthDataType.SLEEP_AWAKE,
+  HealthDataType.SLEEP_DEEP,
+  HealthDataType.SLEEP_LIGHT,
+  HealthDataType.SLEEP_OUT_OF_BED,
+  HealthDataType.SLEEP_REM,
+  HealthDataType.SLEEP_SESSION,
+  HealthDataType.SLEEP_UNKNOWN,
+  HealthDataType.WATER,
+  HealthDataType.WORKOUT,
+  HealthDataType.RESTING_HEART_RATE,
+  HealthDataType.FLIGHTS_CLIMBED,
+  HealthDataType.BASAL_ENERGY_BURNED,
+  HealthDataType.RESPIRATORY_RATE,
+  HealthDataType.NUTRITION,
+  HealthDataType.TOTAL_CALORIES_BURNED,
+  HealthDataType.MENSTRUATION_FLOW,
+];
+
 /// Maps a [HealthDataType] to a [HealthDataUnit].
 const Map<HealthDataType, HealthDataUnit> dataTypeToUnit = {
   HealthDataType.ACTIVE_ENERGY_BURNED: HealthDataUnit.KILOCALORIE,
+  HealthDataType.ATRIAL_FIBRILLATION_BURDEN: HealthDataUnit.PERCENT,
   HealthDataType.AUDIOGRAM: HealthDataUnit.DECIBEL_HEARING_LEVEL,
   HealthDataType.BASAL_ENERGY_BURNED: HealthDataUnit.KILOCALORIE,
   HealthDataType.BLOOD_GLUCOSE: HealthDataUnit.MILLIGRAM_PER_DECILITER,
@@ -117,12 +272,47 @@ const Map<HealthDataType, HealthDataUnit> dataTypeToUnit = {
   HealthDataType.DIETARY_ENERGY_CONSUMED: HealthDataUnit.KILOCALORIE,
   HealthDataType.DIETARY_FATS_CONSUMED: HealthDataUnit.GRAM,
   HealthDataType.DIETARY_PROTEIN_CONSUMED: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_FIBER: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_SUGAR: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_FAT_MONOUNSATURATED: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_FAT_POLYUNSATURATED: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_FAT_SATURATED: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_CHOLESTEROL: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_VITAMIN_A: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_THIAMIN: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_RIBOFLAVIN: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_NIACIN: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_PANTOTHENIC_ACID: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_VITAMIN_B6: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_BIOTIN: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_VITAMIN_B12: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_VITAMIN_C: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_VITAMIN_D: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_VITAMIN_E: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_VITAMIN_K: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_FOLATE: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_CALCIUM: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_CHLORIDE: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_IRON: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_MAGNESIUM: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_PHOSPHORUS: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_POTASSIUM: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_SODIUM: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_ZINC: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_CHROMIUM: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_COPPER: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_IODINE: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_MANGANESE: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_MOLYBDENUM: HealthDataUnit.GRAM,
+  HealthDataType.DIETARY_SELENIUM: HealthDataUnit.GRAM,
+
   HealthDataType.ELECTRODERMAL_ACTIVITY: HealthDataUnit.SIEMEN,
   HealthDataType.FORCED_EXPIRATORY_VOLUME: HealthDataUnit.LITER,
   HealthDataType.HEART_RATE: HealthDataUnit.BEATS_PER_MINUTE,
   HealthDataType.RESPIRATORY_RATE: HealthDataUnit.RESPIRATIONS_PER_MINUTE,
   HealthDataType.PERIPHERAL_PERFUSION_INDEX: HealthDataUnit.PERCENT,
   HealthDataType.HEIGHT: HealthDataUnit.METER,
+  HealthDataType.INSULIN_DELIVERY: HealthDataUnit.INTERNATIONAL_UNIT,
   HealthDataType.RESTING_HEART_RATE: HealthDataUnit.BEATS_PER_MINUTE,
   HealthDataType.STEPS: HealthDataUnit.COUNT,
   HealthDataType.WAIST_CIRCUMFERENCE: HealthDataUnit.METER,
@@ -132,21 +322,19 @@ const Map<HealthDataType, HealthDataUnit> dataTypeToUnit = {
   HealthDataType.DISTANCE_SWIMMING: HealthDataUnit.METER,
   HealthDataType.DISTANCE_CYCLING: HealthDataUnit.METER,
   HealthDataType.FLIGHTS_CLIMBED: HealthDataUnit.COUNT,
-  HealthDataType.MOVE_MINUTES: HealthDataUnit.MINUTE,
   HealthDataType.DISTANCE_DELTA: HealthDataUnit.METER,
 
   HealthDataType.WATER: HealthDataUnit.LITER,
-  HealthDataType.SLEEP_IN_BED: HealthDataUnit.MINUTE,
   HealthDataType.SLEEP_ASLEEP: HealthDataUnit.MINUTE,
-  HealthDataType.SLEEP_ASLEEP_CORE: HealthDataUnit.MINUTE,
-  HealthDataType.SLEEP_ASLEEP_DEEP: HealthDataUnit.MINUTE,
-  HealthDataType.SLEEP_ASLEEP_REM: HealthDataUnit.MINUTE,
   HealthDataType.SLEEP_AWAKE: HealthDataUnit.MINUTE,
+  HealthDataType.SLEEP_AWAKE_IN_BED: HealthDataUnit.MINUTE,
   HealthDataType.SLEEP_DEEP: HealthDataUnit.MINUTE,
-  HealthDataType.SLEEP_REM: HealthDataUnit.MINUTE,
-  HealthDataType.SLEEP_OUT_OF_BED: HealthDataUnit.MINUTE,
+  HealthDataType.SLEEP_IN_BED: HealthDataUnit.MINUTE,
   HealthDataType.SLEEP_LIGHT: HealthDataUnit.MINUTE,
+  HealthDataType.SLEEP_OUT_OF_BED: HealthDataUnit.MINUTE,
+  HealthDataType.SLEEP_REM: HealthDataUnit.MINUTE,
   HealthDataType.SLEEP_SESSION: HealthDataUnit.MINUTE,
+  HealthDataType.SLEEP_UNKNOWN: HealthDataUnit.MINUTE,
 
   HealthDataType.MINDFULNESS: HealthDataUnit.MINUTE,
   HealthDataType.EXERCISE_TIME: HealthDataUnit.MINUTE,
@@ -158,14 +346,23 @@ const Map<HealthDataType, HealthDataUnit> dataTypeToUnit = {
   HealthDataType.HEADACHE_SEVERE: HealthDataUnit.MINUTE,
   HealthDataType.HEADACHE_UNSPECIFIED: HealthDataUnit.MINUTE,
 
+  HealthDataType.GENDER: HealthDataUnit.NO_UNIT,
+  HealthDataType.BIRTH_DATE: HealthDataUnit.NO_UNIT,
+  HealthDataType.BLOOD_TYPE: HealthDataUnit.NO_UNIT,
+
   // Heart Rate events (specific to Apple Watch)
   HealthDataType.HIGH_HEART_RATE_EVENT: HealthDataUnit.NO_UNIT,
   HealthDataType.LOW_HEART_RATE_EVENT: HealthDataUnit.NO_UNIT,
   HealthDataType.IRREGULAR_HEART_RATE_EVENT: HealthDataUnit.NO_UNIT,
   HealthDataType.HEART_RATE_VARIABILITY_SDNN: HealthDataUnit.MILLISECOND,
+  HealthDataType.HEART_RATE_VARIABILITY_RMSSD: HealthDataUnit.MILLISECOND,
   HealthDataType.ELECTROCARDIOGRAM: HealthDataUnit.VOLT,
 
   HealthDataType.NUTRITION: HealthDataUnit.NO_UNIT,
+  HealthDataType.MENSTRUATION_FLOW: HealthDataUnit.NO_UNIT,
+
+  // Health Connect
+  HealthDataType.TOTAL_CALORIES_BURNED: HealthDataUnit.KILOCALORIE,
 
   // JB: Added
   HealthDataType.STAIR_ASCENT_SPEED: HealthDataUnit.METERS_PER_SECOND,
@@ -174,9 +371,6 @@ const Map<HealthDataType, HealthDataUnit> dataTypeToUnit = {
   HealthDataType.WALKING_ASYMMETRY_PERCENTAGE: HealthDataUnit.PERCENT,
   HealthDataType.WALKING_DOUBLE_SUPPORT_PERCENTAGE: HealthDataUnit.PERCENT,
   HealthDataType.WALKING_SPEED: HealthDataUnit.METERS_PER_SECOND,
-
-  // Health Connect
-  HealthDataType.TOTAL_CALORIES_BURNED: HealthDataUnit.KILOCALORIE,
 };
 
 /// Classifications for ECG readings.
@@ -368,26 +562,26 @@ enum HealthWorkoutActivityType {
   // Commented for which platform the type are supported
 
   // Both
+  AMERICAN_FOOTBALL,
   ARCHERY,
+  AUSTRALIAN_FOOTBALL,
   BADMINTON,
   BASEBALL,
   BASKETBALL,
   BIKING, // This also entails the iOS version where it is called CYCLING
   BOXING,
   CRICKET,
+  CROSS_COUNTRY_SKIING,
   CURLING,
+  DOWNHILL_SKIING,
   ELLIPTICAL,
   FENCING,
-  AMERICAN_FOOTBALL,
-  AUSTRALIAN_FOOTBALL,
-  SOCCER,
   GOLF,
   GYMNASTICS,
   HANDBALL,
   HIGH_INTENSITY_INTERVAL_TRAINING,
   HIKING,
   HOCKEY,
-  SKATING,
   JUMP_ROPE,
   KICKBOXING,
   MARTIAL_ARTS,
@@ -397,9 +591,9 @@ enum HealthWorkoutActivityType {
   RUGBY,
   RUNNING,
   SAILING,
-  CROSS_COUNTRY_SKIING,
-  DOWNHILL_SKIING,
+  SKATING,
   SNOWBOARDING,
+  SOCCER,
   SOFTBALL,
   SQUASH,
   STAIR_CLIMBING,
@@ -412,112 +606,63 @@ enum HealthWorkoutActivityType {
   YOGA,
 
   // iOS only
-  BOWLING,
-  CROSS_TRAINING,
-  TRACK_AND_FIELD,
-  DISC_SPORTS,
-  LACROSSE,
-  PREPARATION_AND_RECOVERY,
-  FLEXIBILITY,
-  COOLDOWN,
-  WHEELCHAIR_WALK_PACE,
-  WHEELCHAIR_RUN_PACE,
-  HAND_CYCLING,
-  CORE_TRAINING,
-  FUNCTIONAL_STRENGTH_TRAINING,
-  TRADITIONAL_STRENGTH_TRAINING,
-  MIXED_CARDIO,
-  STAIRS,
-  STEP_TRAINING,
-  FITNESS_GAMING,
   BARRE,
+  BOWLING,
   CARDIO_DANCE,
-  SOCIAL_DANCE,
-  MIND_AND_BODY,
-  PICKLEBALL,
   CLIMBING,
+  COOLDOWN,
+  CORE_TRAINING,
+  CROSS_TRAINING,
+  DISC_SPORTS,
   EQUESTRIAN_SPORTS,
   FISHING,
+  FITNESS_GAMING,
+  FLEXIBILITY,
+  FUNCTIONAL_STRENGTH_TRAINING,
+  HAND_CYCLING,
   HUNTING,
-  PLAY,
-  SNOW_SPORTS,
+  LACROSSE,
+  MIND_AND_BODY,
+  MIXED_CARDIO,
   PADDLE_SPORTS,
-  SURFING_SPORTS,
+  PICKLEBALL,
+  PLAY,
+  PREPARATION_AND_RECOVERY,
+  SNOW_SPORTS,
+  SOCIAL_DANCE,
+  STAIRS,
+  STEP_TRAINING,
+  SURFING,
+  TAI_CHI,
+  TRACK_AND_FIELD,
+  TRADITIONAL_STRENGTH_TRAINING,
   WATER_FITNESS,
   WATER_SPORTS,
-  TAI_CHI,
+  WHEELCHAIR_RUN_PACE,
+  WHEELCHAIR_WALK_PACE,
   WRESTLING,
 
   // Android only
-  AEROBICS,
-  BIATHLON,
-  BIKING_HAND,
-  BIKING_MOUNTAIN,
-  BIKING_ROAD,
-  BIKING_SPINNING,
   BIKING_STATIONARY,
-  BIKING_UTILITY,
   CALISTHENICS,
-  CIRCUIT_TRAINING,
-  CROSS_FIT,
   DANCING,
-  DIVING,
-  ELEVATOR,
-  ERGOMETER,
-  ESCALATOR,
   FRISBEE_DISC,
-  GARDENING,
   GUIDED_BREATHING,
-  HORSEBACK_RIDING,
-  HOUSEWORK,
-  INTERVAL_TRAINING,
-  IN_VEHICLE,
   ICE_SKATING,
-  KAYAKING,
-  KETTLEBELL_TRAINING,
-  KICK_SCOOTER,
-  KITE_SURFING,
-  MEDITATION,
-  MIXED_MARTIAL_ARTS,
-  P90X,
   PARAGLIDING,
-  POLO,
   ROCK_CLIMBING, // on iOS this is the same as CLIMBING
   ROWING_MACHINE,
-  RUNNING_JOGGING, // on iOS this is the same as RUNNING
-  RUNNING_SAND, // on iOS this is the same as RUNNING
   RUNNING_TREADMILL, // on iOS this is the same as RUNNING
   SCUBA_DIVING,
-  SKATING_CROSS, // on iOS this is the same as SKATING
-  SKATING_INDOOR, // on iOS this is the same as SKATING
-  SKATING_INLINE, // on iOS this is the same as SKATING
   SKIING,
-  SKIING_BACK_COUNTRY,
-  SKIING_KITE,
-  SKIING_ROLLER,
-  SLEDDING,
-  SNOWMOBILE,
   SNOWSHOEING,
   STAIR_CLIMBING_MACHINE,
-  STANDUP_PADDLEBOARDING,
-  STILL,
   STRENGTH_TRAINING,
-  SURFING,
   SWIMMING_OPEN_WATER,
   SWIMMING_POOL,
-  TEAM_SPORTS,
-  TILTING,
-  VOLLEYBALL_BEACH,
-  VOLLEYBALL_INDOOR,
-  WAKEBOARDING,
-  WALKING_FITNESS,
-  WALKING_NORDIC,
-  WALKING_STROLLER,
   WALKING_TREADMILL,
   WEIGHTLIFTING,
   WHEELCHAIR,
-  WINDSURFING,
-  ZUMBA,
 
   //
   OTHER,
@@ -529,6 +674,25 @@ enum MealType {
   DINNER,
   SNACK,
   UNKNOWN,
+}
+
+/// Classifications for ECG readings.
+enum ElectrocardiogramClassification {
+  NOT_SET,
+  SINUS_RHYTHM,
+  ATRIAL_FIBRILLATION,
+  INCONCLUSIVE_LOW_HEART_RATE,
+  INCONCLUSIVE_HIGH_HEART_RATE,
+  INCONCLUSIVE_POOR_READING,
+  INCONCLUSIVE_OTHER,
+  UNRECOGNIZED,
+}
+
+/// Types of insulin delivery reason
+enum InsulinDeliveryReason {
+  NOT_SET,
+  BASAL,
+  BOLUS,
 }
 
 /// Extension to assign numbers to [ElectrocardiogramClassification]s
