@@ -1,9 +1,9 @@
-part of mobility_app;
+part of 'main.dart';
 
 class PlacesPage extends StatelessWidget {
   final List<Place> places;
 
-  PlacesPage(this.places);
+  const PlacesPage(this.places, {super.key});
 
   Widget placeEntry(Place p) {
     String lat = p.geoLocation!.latitude.toStringAsFixed(4);
@@ -11,24 +11,18 @@ class PlacesPage extends StatelessWidget {
 
     return Container(
         padding: const EdgeInsets.all(2),
-        margin: EdgeInsets.all(3),
+        margin: const EdgeInsets.all(3),
         child: ListTile(
           leading: Text('Place ID ${p.id}'),
           title: Text('$lat, $lon'),
-          trailing: Text('${formatDuration(p.duration)}'),
+          trailing: Text(formatDuration(p.duration)),
         ));
   }
 
-  Widget list() {
-    return ListView.builder(
-        itemCount: places.length,
-        itemBuilder: (ctx, index) => placeEntry(places[index]));
-  }
+  Widget list() => ListView.builder(
+      itemCount: places.length,
+      itemBuilder: (ctx, index) => placeEntry(places[index]));
 
   @override
-  Widget build(BuildContext context) {
-    return Container(
-      child: list(),
-    );
-  }
+  Widget build(BuildContext context) => Container(child: list());
 }
