@@ -2,17 +2,20 @@ library mobility_features;
 
 import 'dart:async';
 import 'dart:math';
-import 'package:simple_cluster/simple_cluster.dart';
-import 'package:stats/stats.dart';
 import 'dart:core';
 import 'dart:convert';
 import 'dart:io';
-import 'package:path_provider/path_provider.dart';
 
-part 'package:mobility_features/src/mobility_functions.dart';
-part 'package:mobility_features/src/mobility_domain.dart';
-part 'package:mobility_features/src/mobility_intermediate.dart';
-part 'package:mobility_features/src/mobility_context.dart';
-part 'package:mobility_features/src/mobility_serializer.dart';
-part 'package:mobility_features/src/mobility_features.dart';
-part 'package:mobility_features/src/mobility_file_util.dart';
+import 'package:stats/stats.dart';
+import 'package:path_provider/path_provider.dart';
+import 'package:json_annotation/json_annotation.dart';
+import 'package:simple_cluster/simple_cluster.dart';
+import 'package:carp_serializable/carp_serializable.dart';
+
+part 'src/mobility_context.dart';
+part 'src/domain.dart';
+part 'src/main.dart';
+part 'src/util.dart';
+part 'src/mobility_functions.dart';
+part 'src/serializer.dart';
+part 'mobility_features.g.dart';
